Add support for matchDepthView property
Allows WebXR sessions to parse the matchDepthView property from
XrDepthStateInit. This allows platforms with non-view-aligned depth
views (currently just OpenXR for Chrome), to avoid potentially costly
reprojections if the page may not always need them. In order to do this,
information about the depth view thus needs to be exposed to the page.

Bug: 410607163
Change-Id: Id0c5b366201c7ce30f47db7cf648273d98499be1